snackbar en todos los activities listo.

+++ app/src/main/java/com/fernandopretell/rappidemo/util/DownloadFileAsyncTask.java
@@ -1,13 +1,8 @@
 package com.fernandopretell.rappidemo.util;
 
-import android.content.Intent;
-import android.net.Uri;
 import android.os.AsyncTask;
 import android.os.Environment;
 import android.util.Log;
-
-import androidx.core.content.FileProvider;
-
 import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
@@ -36,42 +31,46 @@ public class DownloadFileAsyncTask extends AsyncTask<InputStream, Void, Boolean>
             success = imageRoot.mkdirs();
         }
 
-        if(success){
-            OutputStream output = null;
-            try {
-                output = new FileOutputStream(file);
-
-                byte[] buffer = new byte[1024]; // or other buffer size
-                int read;
-
-                Log.d(TAG, "Attempting to write to: " + imageRoot + "/" + filename);
-                while ((read = inputStream.read(buffer)) != -1) {
-                    output.write(buffer, 0, read);
-                    Log.v(TAG, "Writing to buffer to output stream.");
-                }
-                Log.d(TAG, "Flushing output stream.");
-                output.flush();
-                Log.d(TAG, "Output flushed.");
-            } catch (IOException e) {
-                Log.e(TAG, "IO Exception: " + e.getMessage());
-                e.printStackTrace();
-                return false;
-            } finally {
+        if(!file.exists()){
+            if(success){
+                OutputStream output = null;
                 try {
-                    if (output != null) {
-                        output.close();
-                        Log.d(TAG, "Output stream closed sucessfully.");
+                    output = new FileOutputStream(file);
+
+                    byte[] buffer = new byte[1024]; // or other buffer size
+                    int read;
+
+                    Log.d(TAG, "Attempting to write to: " + imageRoot + "/" + filename);
+                    while ((read = inputStream.read(buffer)) != -1) {
+                        output.write(buffer, 0, read);
+                        Log.v(TAG, "Writing to buffer to output stream.");
                     }
-                    else{
-                        Log.d(TAG, "Output stream is null");
-                    }
-                } catch (IOException e){
-                    Log.e(TAG, "Couldn't close output stream: " + e.getMessage());
+                    Log.d(TAG, "Flushing output stream.");
+                    output.flush();
+                    Log.d(TAG, "Output flushed.");
+                } catch (IOException e) {
+                    Log.e(TAG, "IO Exception: " + e.getMessage());
                     e.printStackTrace();
                     return false;
+                } finally {
+                    try {
+                        if (output != null) {
+                            output.close();
+                            Log.d(TAG, "Output stream closed sucessfully.");
+                        }
+                        else{
+                            Log.d(TAG, "Output stream is null");
+                        }
+                    } catch (IOException e){
+                        Log.e(TAG, "Couldn't close output stream: " + e.getMessage());
+                        e.printStackTrace();
+                        return false;
+                    }
                 }
             }
         }
+
+
 
 
         return true;
